Realigned bounding boxes
When a bounding object is generated, it looks through the given vector
list for a min and a max. It then uses that min and max to get the eight
points in the bounding box in model space. When collisions are checked,
the bounding object transforms those eight vertices to world space then
uses them to recalculate the min and the max, which it uses to check for
collisions.